backend for puzzles

+++ database/migrations/2025_04_03_032501_create_puzzle_attempts_table.php
@@ -5,18 +5,33 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
     public function up(): void
     {
         Schema::create('puzzle_attempts', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id')->default(1); // Testing, later replace with auth user
-            $table->unsignedInteger('puzzle_num');
+            $table->float('puzzle_num');
             $table->string('entered_key');
             $table->boolean('is_correct')->default(false);
             $table->timestamps();
+
+            // $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+
+            // $table->foreign('puzzle_num')->references('puzzle_num')->on('puzzles')->onDelete('cascade');
         });
     }
 
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
     public function down(): void
     {
         Schema::dropIfExists('puzzle_attempts');
